+++ src/app/_models/query-params-keys.ts
@@ -1,4 +1,0 @@
-export const CLIENT_ID_KEY = 'gcClientId';
-export const LANG_KEY = 'gcLangTag';
-export const ENV_KEY = 'gcTargetEnv';
-export const PATH_KEY = 'wcPath';