+++ src/app/_pipes/index.ts
@@ -1,2 +1,0 @@
-export * from './status-name.pipe';
-export * from './custom-time.pipe';