+++ src/app/_services/index.ts
@@ -1,8 +1,0 @@
-export * from './auth.service';
-export * from './clock.service';
-export * from './agents.service';
-export * from './queues.service';
-export * from './api/notification-api.service';
-export * from './api/presence-api.service';
-export * from './api/users-api.service';
-