+++ src/app/_shared/components/not-found/not-found.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'app-not-found',
+  templateUrl: './not-found.component.html',
+  imports: [
+    CommonModule,
+  ],
+  providers:[]
+})
+export class NotFoundComponent {}