+++ src/app/_shared/models/query-params-keys.ts
@@ -1,0 +1,6 @@
+export const CLIENT_ID_KEY = 'gcClientId';
+export const LANG_KEY = 'gcLangTag';
+export const ENV_KEY = 'gcTargetEnv';
+export const PATH_KEY = 'wcPath';
+export const CONVERSATION_KEY = 'gcConversationId';
+export const HOST_ORIGIN_KEY = 'gcHostOrigin';