+++ src/app/dashboard/components/queues/queues.component.ts
@@ -1,0 +1,67 @@
+import { Component, inject, input, model, output } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { TableModule } from 'primeng/table';
+import { PanelStatusEnum } from '../../models';
+import { Button } from 'primeng/button';
+import { Models } from 'purecloud-platform-client-v2';
+import { FilterMetadata } from 'primeng/api';
+import { Tooltip } from 'primeng/tooltip';
+import { TranslatePipe } from '@ngx-translate/core';
+import { CustomTimePipe, PhoneFormatPipe } from '../../pipes';
+import { QueuesService } from '../../services';
+import { ClockService } from '../../../_shared/services';
+
+@Component({
+  selector: 'app-queues',
+  templateUrl: './queues.component.html',
+  styleUrl: './queues.component.scss',
+  imports: [
+    CommonModule,
+    TableModule,
+    CustomTimePipe,
+    Button,
+    PhoneFormatPipe,
+    Tooltip,
+    TranslatePipe
+  ]
+})
+export class QueuesComponent {
+  status = input.required<PanelStatusEnum>();
+  public readonly queuesSvc = inject(QueuesService);
+  public readonly clockSvc = inject(ClockService);
+  toggleStatus = output();
+  filters = model.required<{
+    [s: string]: FilterMetadata | FilterMetadata[];
+  }>();
+  sortOrder = model.required<number>();
+
+
+  protected readonly PanelStatusEnum = PanelStatusEnum;
+  private readonly service = inject(QueuesService);
+
+  onFilter(event: any) {
+    if (JSON.stringify(this.filters()) !== JSON.stringify(event.filters)) {
+      this.filters.set(event.filters);
+    }
+  }
+
+  onSort(event: any) {
+    if (this.sortOrder() !== event.order) {
+      this.sortOrder.set(event.order);
+    }
+  }
+
+  hasPermission(mediaType: string): boolean {
+    return this.service.hasPermission(mediaType);
+  }
+
+  isOnQueue(): boolean {
+    return this.service.isOnQueue();
+  }
+
+  assignToMe(interaction: Models.ConversationActivityEntityData) {
+    if (!interaction.conversationId) return;
+    this.service.assignInteraction(interaction.conversationId);
+  }
+
+}