+++ src/app/dashboard/pipes/index.ts
@@ -1,0 +1,5 @@
+export * from './custom-time.pipe';
+export * from './phone-format.pipe';
+export * from './queue-name.pipe';
+export * from './status-name.pipe';
+