+++ src/app/email-widget/components/home/home.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+
+@Component({
+  selector: 'app-email-widget-home',
+  templateUrl: './home.component.html',
+  imports: [
+    CommonModule,
+  ],
+  providers:[]
+})
+export class HomeComponent {}